JAMES-1852 Move mailets to the right package
It allows it to be addressed in mailetcontainer by name and not by FQDN
